Renaming and file reordering
