+++ src/Nethermind/Nethermind.Optimism/Rpc/IOptimismEngineRpcModule.cs
@@ -6,6 +6,7 @@ using Nethermind.Core.Crypto;
 using Nethermind.JsonRpc;
 using Nethermind.JsonRpc.Modules;
 using Nethermind.Merge.Plugin.Data;
+using Nethermind.Optimism.ProtocolVersion;
 
 namespace Nethermind.Optimism.Rpc;
